move mongorest class out of __init__
